--- android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/around/AroundUiState.kt
@@ -1,10 +1,0 @@
-package com.boostcampwm2023.snappoint.presentation.around
-
-import com.boostcampwm2023.snappoint.presentation.model.PostSummaryState
-
-data class AroundUiState(
-    val posts: List<PostSummaryState> = emptyList(),
-    val onPreviewButtonClicked: (Int) -> Unit,
-    val onViewPostButtonClicked: (Int) -> Unit,
-)
-
