+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/auth/signup/SignupEvent.kt
@@ -1,0 +1,6 @@
+package com.boostcampwm2023.snappoint.presentation.auth.signup
+
+sealed class SignupEvent {
+    data object NavigateToSignIn : SignupEvent()
+    data class ShowMessage(val messageResId: Int) : SignupEvent()
+}
