+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/auth/signup/SignupUiState.kt
@@ -1,0 +1,14 @@
+package com.boostcampwm2023.snappoint.presentation.auth.signup
+
+data class SignupUiState(
+    val email: String = "",
+    val password: String = "",
+    val passwordConfirm: String = "",
+    val nickname: String = "",
+    val isSignUpInProgress: Boolean = false,
+    val isButtonEnabled: Boolean = false,
+    val emailErrorResId: Int? = null,
+    val passwordErrorResId: Int? = null,
+    val passwordConfirmErrorResId: Int? = null,
+    val nicknameErrorResId: Int? = null
+)
