+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/around/AroundEvent.kt
@@ -1,0 +1,6 @@
+package com.boostcampwm2023.snappoint.presentation.main.around
+
+sealed class AroundEvent {
+    data class ShowSnapPointAndRoute(val index: Int): AroundEvent()
+    data class NavigateViewPost(val index: Int): AroundEvent()
+}
