+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/around/AroundUiState.kt
@@ -1,0 +1,10 @@
+package com.boostcampwm2023.snappoint.presentation.main.around
+
+import com.boostcampwm2023.snappoint.presentation.model.PostSummaryState
+
+data class AroundUiState(
+    val posts: List<PostSummaryState> = emptyList(),
+    val onPreviewButtonClicked: (Int) -> Unit,
+    val onViewPostButtonClicked: (Int) -> Unit,
+)
+
