+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/popularpost/PopularPostFragment.kt
@@ -1,0 +1,9 @@
+package com.boostcampwm2023.snappoint.presentation.main.popularpost
+
+import androidx.fragment.app.Fragment
+import com.boostcampwm2023.snappoint.R
+import com.boostcampwm2023.snappoint.databinding.FragmentPopularPostBinding
+import com.boostcampwm2023.snappoint.presentation.base.BaseFragment
+
+class PopularPostFragment : BaseFragment<FragmentPopularPostBinding>(R.layout.fragment_popular_post) {
+}
