+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/preview/PreviewUiState.kt
@@ -1,0 +1,9 @@
+package com.boostcampwm2023.snappoint.presentation.main.preview
+
+import com.boostcampwm2023.snappoint.presentation.model.PostBlockState
+
+data class PreviewUiState(
+    val title: String = "",
+    val timeStamp: String = "",
+    val blocks: List<PostBlockState> = emptyList()
+)
