+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/preview/PreviewViewHolder.kt
@@ -1,0 +1,24 @@
+package com.boostcampwm2023.snappoint.presentation.main.preview
+
+import android.net.Uri
+import android.util.Log
+import androidx.recyclerview.widget.RecyclerView
+import coil.load
+import coil.request.CachePolicy
+import com.boostcampwm2023.snappoint.databinding.ItemImagePreviewBinding
+
+class PreviewViewHolder(
+    private val binding: ItemImagePreviewBinding
+) : RecyclerView.ViewHolder(binding.root) {
+
+    fun bind(uri: Uri, description: String) {
+        with(binding){
+            ivPreviewCarouselImage.load(uri) {
+                memoryCachePolicy(CachePolicy.ENABLED)
+                diskCachePolicy(CachePolicy.ENABLED)
+                scale(coil.size.Scale.FILL)
+            }
+            tvPreviewImageDescription.text = description
+        }
+    }
+}
