+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/setting/SettingEvent.kt
@@ -1,0 +1,7 @@
+package com.boostcampwm2023.snappoint.presentation.main.setting
+
+sealed class SettingEvent {
+    data object RemoveSnapPoint : SettingEvent()
+    data object SignOut : SettingEvent()
+    data object FailToSignOut: SettingEvent()
+}
