--- android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/popularpost/PopularPostFragment.kt
@@ -1,9 +1,0 @@
-package com.boostcampwm2023.snappoint.presentation.popularpost
-
-import androidx.fragment.app.Fragment
-import com.boostcampwm2023.snappoint.R
-import com.boostcampwm2023.snappoint.databinding.FragmentPopularPostBinding
-import com.boostcampwm2023.snappoint.presentation.base.BaseFragment
-
-class PopularPostFragment : BaseFragment<FragmentPopularPostBinding>(R.layout.fragment_popular_post) {
-}
