--- android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/setting/SettingEvent.kt
@@ -1,7 +1,0 @@
-package com.boostcampwm2023.snappoint.presentation.setting
-
-sealed class SettingEvent {
-    data object RemoveSnapPoint : SettingEvent()
-    data object SignOut : SettingEvent()
-    data object FailToSignOut: SettingEvent()
-}
