--- android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/signin/SignInFormState.kt
@@ -1,9 +1,0 @@
-package com.boostcampwm2023.snappoint.presentation.signin
-
-data class SignInFormState(
-    val email: String = "",
-    val password: String = "",
-    val isEmailValid: Boolean = false,
-    val isPasswordValid: Boolean = false,
-    val isLoginInProgress: Boolean = false
-)
