--- android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/signup/SignupEvent.kt
@@ -1,6 +1,0 @@
-package com.boostcampwm2023.snappoint.presentation.signup
-
-sealed class SignupEvent {
-    data object NavigateToSignIn : SignupEvent()
-    data class ShowMessage(val messageResId: Int) : SignupEvent()
-}
