+++ android/app/src/main/res/layout/fragment_popular_post.xml
@@ -9,7 +9,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.popularpost.PopularPostFragment">
+        tools:context=".presentation.main.popularpost.PopularPostFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
