+++ android/app/src/main/res/layout/fragment_preview.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.boostcampwm2023.snappoint.presentation.preview.PreviewViewModel" />
+            type="com.boostcampwm2023.snappoint.presentation.main.preview.PreviewViewModel" />
 
         <import type="android.view.View" />
     </data>
@@ -15,7 +15,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.preview.PreviewFragment">
+        tools:context=".presentation.main.preview.PreviewFragment">
 
         <ImageView
             android:id="@+id/iv_preview_profile"
