+++ android/app/src/main/res/layout/fragment_subscription.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.subscription.SubscriptionFragment">
+        tools:context=".presentation.main.subscription.SubscriptionFragment">
 
         <!-- TODO: Update blank fragment layout -->
