+++ android/app/src/main/res/navigation/nav_auth.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/signInFragment"
-        android:name="com.boostcampwm2023.snappoint.presentation.signin.SignInFragment"
+        android:name="com.boostcampwm2023.snappoint.presentation.auth.signin.SignInFragment"
         android:label="LoginFragment" >
         <action
             android:id="@+id/action_loginFragment_to_signupFragment"
@@ -18,7 +18,7 @@
     </fragment>
     <fragment
         android:id="@+id/signUpFragment"
-        android:name="com.boostcampwm2023.snappoint.presentation.signup.SignupFragment"
+        android:name="com.boostcampwm2023.snappoint.presentation.auth.signup.SignupFragment"
         android:label="SignupFragment" />
     <activity
         android:id="@+id/mainActivity"
